move unrelated files into more related FS
